Move the test out into tests/ as it covers end to end.
`server_test.go` should IMO only cover the unit aspects of everything
that is in `server.go`
