+++ tests/server_test.go
@@ -1,4 +1,4 @@
-package daemon
+package tests
 
 import (
 	"bytes"
@@ -11,24 +11,25 @@ import (
 	"testing"
 
 	"github.com/Bitspark/slang/pkg/core"
+	"github.com/Bitspark/slang/pkg/daemon"
 	"github.com/Bitspark/slang/pkg/env"
 	"github.com/Bitspark/slang/pkg/storage"
 	"github.com/Bitspark/slang/tests"
 	"github.com/stretchr/testify/assert"
 )
 
-func getTestServer() *Server {
+func getTestServer() *daemon.Server {
 	backend := tests.NewTestLoader("./")
 	env := env.New("localhost", 8000)
 	storage := storage.NewStorage().AddBackend(backend)
-	ctx := SetStorage(context.Background(), storage)
+	ctx := daemon.SetStorage(context.Background(), storage)
 	backend.Reload()
 	fmt.Println(backend.List())
-	return NewServer(&ctx, env)
+	return daemon.NewServer(&ctx, env)
 }
 func TestServer_operator_starting(t *testing.T) {
 	server := getTestServer()
-	data := runInstructionJSON{Id: "8b62495a-e482-4a3e-8020-0ab8a350ad2d",
+	data := daemon.RunInstructionJSON{Id: "8b62495a-e482-4a3e-8020-0ab8a350ad2d",
 		Stream: false,
 		Props:  core.Properties{"value": "slang"},
 		Gens: core.Generics{
@@ -42,7 +43,7 @@ func TestServer_operator_starting(t *testing.T) {
 	response := httptest.NewRecorder()
 	server.Handler().ServeHTTP(response, request)
 	decoder := json.NewDecoder(response.Body)
-	var out outJSON
+	var out daemon.InstanceStateJSON
 	decoder.Decode(&out)
 	assert.Equal(t, 200, response.Code)
 	assert.Equal(t, "success", out.Status)
